+++ fvm/utils/cadenceValues.go
@@ -42,8 +42,8 @@ func CadenceValueToAddressSlice(value cadence.Value) (addresses []common.Address
 	return addresses, true
 }
 
-func CadenceValueToUintUintMap(value cadence.Value) (map[uint]uint, bool) {
-	result := make(map[uint]uint)
+func CadenceValueToUintUintMap(value cadence.Value) (map[uint]uint64, bool) {
+	result := make(map[uint]uint64)
 
 	dict, ok := value.(cadence.Dictionary)
 	if !ok {
@@ -51,17 +51,17 @@ func CadenceValueToUintUintMap(value cadence.Value) (map[uint]uint, bool) {
 	}
 
 	for _, p := range dict.Pairs {
-		key, ok := p.Key.(cadence.UInt32)
+		key, ok := p.Key.(cadence.UInt64)
 		if !ok {
 			return nil, false
 		}
 
-		value, ok := p.Value.(cadence.UInt32)
+		value, ok := p.Value.(cadence.UInt64)
 		if !ok {
 			return nil, false
 		}
 
-		result[uint(key)] = uint(value)
+		result[uint(key)] = uint64(value)
 	}
 
 	return result, true
